requirements.txt
numpy==1.8.0
# requests==2.5.0
# beautifulsoup4==4.3.2
twython==3.2.0
pyPNG==0.0.17
translate==0.0.6
